requirements.txt
torch==1.13.0
torchvision==0.14.0
numpy==1.23.4
tensorboardx==2.5.1
matplotlib==3.6.2
tqdm==4.64.1
imutils
pandas==1.5.1
tensorboard
Django==3.2.12
mysqlclient==2.1.0
djangorestframework==3.13.1